Добавлен User Controller 24.04.2023 К нему сервис и репозиторий

+++ src/main/java/team/skyprojava/websitebackend/repository/UserRepository.java
@@ -6,8 +6,11 @@ import team.skyprojava.websitebackend.entity.User;
 
 import java.util.Optional;
 
+/**
+ * Repository UserRepository (users/пользователь).
+ */
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
